--- src/main/kotlin/index/util/KeyTool.kt
@@ -1,197 +1,0 @@
-package index.util
-
-import index.btree.logger
-import mu.KotlinLogging
-import java.lang.IndexOutOfBoundsException
-import java.nio.ByteBuffer
-import java.nio.charset.StandardCharsets
-import java.time.Instant
-import java.time.LocalDate
-import java.time.LocalDateTime
-import java.time.ZoneOffset
-import java.util.UUID
-
-/*
-* page 로 넘어가게 되면 page 크기를 기준으로 고정 크기 buffer 사용 할 것
-* */
-object KeyTool {
-    private val log = KotlinLogging.logger {}
-
-    fun pack(keyList: List<Any?>, schema: KeySchema): ByteArray{
-        require(keyList.size <= schema.columns.size) { "Too many key values for schema" }
-        return buildList{
-            for (idx in keyList.indices){
-                val packed = packKeyItem(keyList[idx], schema.columns[idx])
-                addAll(packed.toList())
-            }
-        }.toByteArray()
-    }
-
-    fun unpack(bytes: ByteArray, schema: KeySchema): List<Any?>{
-        val unpackedKeys = mutableListOf<Any?>()
-        var offset = 0
-
-        for (column in schema.columns){
-            val (value, consumed) = try {
-                unpackKeyItem(bytes, offset, column)
-            } catch (_: IndexOutOfBoundsException){
-                break
-            }
-            unpackedKeys.add(value)
-            offset += consumed
-        }
-        return unpackedKeys
-    }
-
-    private fun packKeyItem(key: Any?, column: Column): ByteArray{
-        if(key == null) return byteArrayOf(0x00)
-        return when (column.type){
-            ColumnType.BOOLEAN -> {
-                val packedKey  = byteArrayOf(if (key as Boolean) 1 else 0)
-                (byteArrayOf(0x01)) + packedKey
-            }
-
-            ColumnType.BYTE -> {
-                val packedKey = byteArrayOf(key as Byte)
-                (byteArrayOf(0x01)) + packedKey
-            }
-
-            ColumnType.SHORT -> {
-                val packedKey = (key as Short).encodeSortable()
-                (byteArrayOf(0x01)) + packedKey
-            }
-            ColumnType.INT -> {
-                val packedKey = (key as Int).encodeSortable()
-                (byteArrayOf(0x01)) + packedKey
-            }
-            ColumnType.LONG -> {
-                val packedKey = (key as Long).encodeSortable()
-                (byteArrayOf(0x01)) + packedKey
-            }
-            ColumnType.FLOAT -> {
-                val packedKey = (key as Float).encodeSortable()
-                (byteArrayOf(0x01)) + packedKey
-            }
-            ColumnType.DOUBLE -> {
-                val packedKey = (key as Double).encodeSortable()
-                (byteArrayOf(0x01)) + packedKey
-            }
-            ColumnType.STRING -> {
-                val rawString = key as String
-                val bytes = column.collation?.getCollationKey(rawString)?.toByteArray() ?: rawString.toByteArray(StandardCharsets.UTF_8)
-                (byteArrayOf(0x01)) + encodeVarInt(bytes.size) + bytes
-            }
-
-            ColumnType.LOCAL_DATE -> {
-                val epochDay = (key as LocalDate).toEpochDay()
-                val packedKey = epochDay.encodeSortable()
-                (byteArrayOf(0x01)) + packedKey
-            }
-            ColumnType.LOCAL_DATE_TIME -> {
-                val epochSecond = (key as LocalDateTime).toEpochSecond(ZoneOffset.UTC)
-                val packedKey = epochSecond.encodeSortable()
-                (byteArrayOf(0x01)) + packedKey
-            }
-            ColumnType.INSTANT -> {
-                val epochSecond = (key as Instant).epochSecond
-                val packedKey = epochSecond.encodeSortable()
-                (byteArrayOf(0x01)) + packedKey
-            }
-            ColumnType.UUID -> {
-                val packedKey = ByteBuffer.allocate(16)
-                    .putLong((key as UUID).mostSignificantBits)
-                    .putLong(key.leastSignificantBits)
-                    .array()
-                (byteArrayOf(0x01)) + packedKey
-            }
-            ColumnType.BYTES -> {
-                val bytes = key as ByteArray
-                (byteArrayOf(0x01)) + encodeVarInt(bytes.size) + bytes
-            }
-        }
-    }
-
-    private fun unpackKeyItem(bytes: ByteArray, offset: Int, column: Column): Pair<Any?, Int> {
-        var position = offset
-        val nullFlag = try { bytes[position++] } catch( exception: IndexOutOfBoundsException ) { throw exception}
-        if (nullFlag.toInt() == 0x00) return null to 1
-
-        val descending = column.descending
-        val columnType = column.type
-
-        fun readBytes(length: Int): ByteArray {
-            val slice = bytes.copyOfRange(position, position + length)
-            position += length
-            return slice.invert(descending)
-        }
-
-        val result: Any? = when (columnType){
-            ColumnType.BOOLEAN -> bytes[position++].toInt() != 0
-            ColumnType.BYTE -> bytes[position++]
-            ColumnType.SHORT -> {
-                val array = readBytes(2)
-                array.decodeSortableShort()
-            }
-            ColumnType.INT -> {
-                val array = readBytes(4)
-                array.decodeSortableInt()
-            }
-            ColumnType.LONG -> {
-                val array = readBytes(8)
-                array.decodeSortableLong()
-            }
-            ColumnType.FLOAT -> {
-                val array = readBytes(4)
-                array.decodeSortableFloat()
-            }
-            ColumnType.DOUBLE -> {
-                val array = readBytes(8)
-                array.decodeSortableDouble()
-            }
-
-            ColumnType.STRING -> {
-                val (len, lenBytes) = decodeVarInt(bytes, position)
-                position += lenBytes
-                val strBytes = readBytes(len)
-
-                if (column.collation != null){
-                    val result = "[CollationKey(${strBytes.joinToString(" ")})]"
-                    logger.info { "UnpackKey: $result" }
-                    result
-                } else{
-                    strBytes.toString(StandardCharsets.UTF_8)
-                }
-            }
-
-            ColumnType.LOCAL_DATE -> {
-                val array = readBytes(8)
-                val epochDay = array.decodeSortableLong()
-                LocalDate.ofEpochDay(epochDay)
-            }
-
-            ColumnType.LOCAL_DATE_TIME -> {
-                val array = readBytes(8)
-                val epochSecond = array.decodeSortableLong()
-                LocalDateTime.ofEpochSecond(epochSecond, 0, ZoneOffset.UTC)
-            }
-
-            ColumnType.INSTANT -> {
-                val array = readBytes(8)
-                val epochSecond = array.decodeSortableLong()
-                Instant.ofEpochSecond(epochSecond)
-            }
-
-            ColumnType.UUID -> {
-                val array = readBytes(16)
-                val bb = ByteBuffer.wrap(array)
-                UUID(bb.long, bb.long)
-            }
-            ColumnType.BYTES -> {
-                val (len, lenBytes) = decodeVarInt(bytes, position)
-                position += lenBytes
-                readBytes(len)
-            }
-        }
-        return result to (position - offset)
-    }
-}
